Fixed open in browser and periods in file names (but not really).
Open in browser didn't work since mhtml files don't always have the content-url on the same line.

If there were any periods in the filename, the user couldn't open it from a file manager. Now, it'll work as long as there are no more than 7 periods in the filename; this is really dumb but I couldn't find another way to do it.