Added Local Link Setting
Now the user can control what happens when they click on a link in a local file. There are three options:

Ask User:  this will open an alert that will let the user choose between opening a local file or the online url if a local file exists. If the local file doesn't exists, it will still open an alert asking if the user wants to open the online version.

Local Only: this will open the local file corresponding to the link if it exists. If there is no local file for the link, the link doesn't go anywhere.

Do Nothing: this stops all links from working in a local file.