Added FileBrowser Function
Added a function that will call open the file browser dialog at the path of the first parameter and executes the functions of the second parameter when OK or Cancel is selected in the dialog. This should make it possible for me to open a file browser and do other things with the result if that function is ever required anywhere else.